note templates can be stored in bundle files now
