+++ DLPrototype/Views/Entities/Notes/Templates/List.md
